requirements.txt
requests==2.31.0
pydantic==2.5.0
pytest==7.4.0
python-dotenv
black
